Update device name only if it is empty
This commit modifies the `identifyDevice` method in `AdbManager.kt` to ensure that an existing device name is preserved if it is not empty. The device's name will only be updated to a combination of brand and model if the existing `device.name` is empty and both fetched brand and model are not empty.

Key changes:
- In `AdbManager.identifyDevice`, the logic for setting the `name` property of the `Device` has been changed.
- If `device.name` is not empty, it will be retained.
- If `device.name` is empty, it will be set to `"$deviceBrand $deviceName".trim()` if this combined string is not empty.